+++ compiler/lib/v1/compiler.dart
@@ -5,6 +5,8 @@ export 'src/kelicap_compiler/analyzer.dart'
         linkToReference,
         linkTypeOf,
         typeArgumentOf,
+        unerasedTypeOf,
+        unerasedTypeValueOf,
         urlOf,
         DependencyElement,
         DependencyInvocation,