+++ compiler/CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 
+## 1.0.2
+
+- fix: Misc. compile and runtime errors contributed by the community
+
 ## 1.0.1
 
 - fix: Replaced deprecated methods used to get type string in di module
 
 ## 1.0.0
 
-- Initial release
+- chore: Initial release