+++ forms/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.0.2
+
+- fix: Misc. runtime errors contributed by the community
+
 ## 1.0.1
 
 - chore: Update `pubspec.yaml` description